Transliteral conversion of the SimplifyMesh routine to Mastodon data structure.
Ok so I took some time to build the thread leading to this piece
of code.

The paper that started this is the seminal article by Garland
and Heckbert on quadratic mesh decimation:
http://www.cs.cmu.edu/~./garland/Papers/quadrics.pdf (1995)

But there are much more info in Garland's PhD:
https://mgarland.org/files/papers/thesis-onscreen.pdf

Then, Sven Forstmann proposed an implementation of this with speed
improvement called "Fast Quadric Mesh Simplification".
The speed improvement relies on avoiding sorting triangles to
know which to delete.
The C++ code repo is here (MIT license):
https://github.com/sp4cerat/Fast-Quadric-Mesh-Simplification

The simplify function specifically:
https://github.com/sp4cerat/Fast-Quadric-Mesh-Simplification/blob/master/src.cmd/Simplify.h

But there is also a Javascript version here, that inspired the next
person listed here.

Then JayFella ported it to Java, using the JMonkeyEngine data structure.
Discussed here:
https://hub.jmonkeyengine.org/t/isosurface-mesh-simplifier/41046/1

The code by JayFella (MIT license):
https://gist.github.com/jayfella/00a328b2dbdf6304078a821143b7aef7

It was then ported by Deborah Schmidt to use the ImageJ-Mesh
data structure instead of JMesh.

Then finally I (JY) simply used a mastodon TriMesh instead.